requirements.txt
asgiref==3.7.2
boto3==1.28.27
botocore==1.31.27
cffi==1.15.1
Django==4.2.4
django-cors-headers==4.2.0
django-storages==1.13.2
djangorestframework==3.14.0
gunicorn==20.0.4
jmespath==1.0.1
mysql-connector-python==8.1.0
olefile==0.46
Pillow==9.1.0
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
validate-email==1.3